chore: Move auth fixture to shared folder
