+++ tests/api/articles/articles-id-patch.spec.ts
@@ -1,11 +1,11 @@
 import { HttpStatusCode } from '@_src_api/enums/api-status-code.enum';
 import { UserType } from '@_src_api/enums/user-types.enum';
-import { testUsers } from '@_src_fixtures_api/auth';
 import {
   createHeaders,
   createInvalidHeaders,
 } from '@_src_helpers_api/create-token.helper';
 import { APIResponse, expect, test } from '@playwright/test';
+import { testUsers } from 'src/shared/fixtures/auth';
 
 test.describe('PATCH articles/{id} endpoint tests', async () => {
   const articles: string = `/api/articles`;
